[GEOT-6495] Cleanup after merge of ElasticGeo into gs-web-elasticsearch
