+++ doc/en/user/source/community/index.rst
@@ -29,6 +29,7 @@ officially part of the GeoServer releases. They are however built along with the
    wps-download/index
    jms-cluster/index
    solr/index
+   elasticsearch/index
    geomesa/index
    gwc-distributed/index
    flatgeobuf/index
