--- src/community/elasticsearch/src/main/java/mil/nga/giat/elasticsearch/ElasticXStreamPersisterInitializer.java
@@ -1,15 +1,0 @@
-package mil.nga.giat.elasticsearch;
-
-import org.geoserver.config.util.XStreamPersister;
-import org.geoserver.config.util.XStreamPersisterInitializer;
-
-class ElasticXStreamPersisterInitializer implements XStreamPersisterInitializer {
-
-    @Override
-    public void init(XStreamPersister persister) {
-        persister.getXStream().allowTypes(new String[] {
-                "mil.nga.giat.data.elasticsearch.ElasticAttribute"
-        });
-    }
-
-}
