+++ src/community/elasticsearch/src/main/java/org/geoserver/elasticsearch/ElasticConfigurationPanel.html
