+++ src/community/elasticsearch/src/main/java/org/geoserver/elasticsearch/ElasticConfigurationPanelInfo.java
@@ -3,21 +3,20 @@
  * application directory.
  */
 
-package mil.nga.giat.elasticsearch;
+package org.geoserver.elasticsearch;
 
 import org.geoserver.catalog.FeatureTypeInfo;
 import org.geoserver.platform.ExtensionPriority;
 import org.geoserver.web.data.resource.ResourceConfigurationPanelInfo;
 
 /**
- * 
- * Implements ResourceConfigurationPanelInfo extension point to add Elasticsearch 
- * attribute configuration link on resource page. <br>
- * Priority is reduced under standard {@link ExtensionPriority#LOWEST} to shows
- * the Elasticsearch link after other panels.
- * 
+ * Implements ResourceConfigurationPanelInfo extension point to add Elasticsearch attribute
+ * configuration link on resource page. <br>
+ * Priority is reduced under standard {@link ExtensionPriority#LOWEST} to shows the Elasticsearch
+ * link after other panels.
  */
-class ElasticConfigurationPanelInfo extends ResourceConfigurationPanelInfo implements ExtensionPriority {
+class ElasticConfigurationPanelInfo extends ResourceConfigurationPanelInfo
+        implements ExtensionPriority {
 
     private static final long serialVersionUID = 1485404586629946126L;
 
@@ -40,5 +39,4 @@ class ElasticConfigurationPanelInfo extends ResourceConfigurationPanelInfo imple
     public int getPriority() {
         return ExtensionPriority.LOWEST + 1;
     }
-
 }
