+++ src/community/elasticsearch/src/main/java/org/geoserver/elasticsearch/ElasticXStreamInitializer.java
@@ -4,28 +4,25 @@
  * application directory.
  */
 
-package mil.nga.giat.elasticsearch;
-
-import mil.nga.giat.data.elasticsearch.ElasticAttribute;
-import mil.nga.giat.data.elasticsearch.ElasticLayerConfiguration;
-
-import org.geoserver.config.util.XStreamPersister;
-import org.geoserver.config.util.XStreamPersisterInitializer;
+package org.geoserver.elasticsearch;
 
 import com.thoughtworks.xstream.XStream;
+import org.geoserver.config.util.XStreamPersister;
+import org.geoserver.config.util.XStreamPersisterInitializer;
+import org.geotools.data.elasticsearch.ElasticAttribute;
+import org.geotools.data.elasticsearch.ElasticLayerConfiguration;
 
 /**
- * 
- * Implementation of XStreamPersisterInitializer extension point to serialize ElasticLayerConfiguration
- *
+ * Implementation of XStreamPersisterInitializer extension point to serialize
+ * ElasticLayerConfiguration
  */
 class ElasticXStreamInitializer implements XStreamPersisterInitializer {
 
     @Override
     public void init(XStreamPersister persister) {
-        persister.registerBreifMapComplexType("elasticLayerConfiguration",ElasticLayerConfiguration.class);
+        persister.registerBreifMapComplexType(
+                "elasticLayerConfiguration", ElasticLayerConfiguration.class);
         XStream xs = persister.getXStream();
         xs.alias("esAttribute", ElasticAttribute.class);
     }
-
 }
