+++ src/community/elasticsearch/src/main/java/org/geoserver/elasticsearch/ElasticXStreamPersisterInitializer.java
@@ -1,0 +1,18 @@
+/* (c) 2016 Open Source Geospatial Foundation - all rights reserved
+ * This code is licensed under the GPL 2.0 license, available at the root
+ * application directory.
+ */
+package org.geoserver.elasticsearch;
+
+import org.geoserver.config.util.XStreamPersister;
+import org.geoserver.config.util.XStreamPersisterInitializer;
+
+class ElasticXStreamPersisterInitializer implements XStreamPersisterInitializer {
+
+    @Override
+    public void init(XStreamPersister persister) {
+        persister
+                .getXStream()
+                .allowTypes(new String[] {"org.geotools.data.elasticsearch.ElasticAttribute"});
+    }
+}
