+++ src/community/pom.xml
@@ -58,6 +58,7 @@
           <descriptor>release/ext-hz-cluster.xml</descriptor>
           <descriptor>release/ext-activeMQ.xml</descriptor>
           <descriptor>release/ext-solr.xml</descriptor>
+          <descriptor>release/ext-elasticsearch.xml</descriptor>
           <!-- If you are re-enabling this, remember to re-add the documentation -->
           <!-- <descriptor>release/ext-rest-upload.xml</descriptor> -->
           <descriptor>release/ext-gwc-distributed.xml</descriptor>
@@ -245,6 +246,7 @@
         <module>jms-cluster</module>
         <module>hz-cluster</module>
         <module>solr</module>
+        <module>elasticsearch</module>
         <!-- <module>rest-upload</module> -->
         <module>params-extractor</module>
         <module>wps-remote</module>
@@ -403,6 +405,12 @@
       <id>solr</id>
       <modules>
         <module>solr</module>
+      </modules>
+    </profile>
+    <profile>
+      <id>elasticsearch</id>
+      <modules>
+        <module>elasticsearch</module>
       </modules>
     </profile>
     <profile>
