+++ src/community/release/ext-elasticsearch.xml
@@ -1,0 +1,38 @@
+<assembly>
+  <id>elasticsearch-plugin</id>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <includeBaseDirectory>false</includeBaseDirectory>
+  <fileSets>
+    <fileSet>
+      <directory>release/target/dependency</directory>
+      <outputDirectory></outputDirectory>
+      <includes>
+        <include>commons-math3*</include>
+        <include>elasticsearch-rest-client*</include>
+        <include>geo-*</include>
+        <include>grumpy-core*</include>
+        <include>gs-web-elasticsearch*</include>
+        <include>gt-elasticsearch*</include>
+        <include>httpcore-nio*</include>
+        <include>httpasyncclient*</include>
+        <include>httpcore-nio*</include>
+        <include>jackson-databind*</include>
+        <include>jackson-annotations*</include>
+        <include>spatial4j*</include>
+      </includes>
+      <excludes>
+        <exclude>commons-codec*</exclude>
+        <exclude>commons-collections*</exclude>
+        <exclude>commons-httpclient*</exclude>
+        <exclude>commons-logging*</exclude>
+        <exclude>guava*</exclude>
+        <exclude>jai*</exclude>
+        <exclude>jts*</exclude>
+        <exclude>jsr305*</exclude>
+        <exclude>log4j*</exclude>
+      </excludes>
+    </fileSet>
+  </fileSets>
+</assembly>
