+++ src/community/release/pom.xml
@@ -132,6 +132,11 @@
      <version>${project.version}</version>
    </dependency>
    <dependency>
+     <groupId>org.geoserver.community</groupId>
+     <artifactId>gs-web-elasticsearch</artifactId>
+     <version>${project.version}</version>
+   </dependency>
+   <dependency>
       <groupId>org.geoserver.community</groupId>
       <artifactId>gs-gwc-distributed</artifactId>
      <version>${project.version}</version>
